+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.20.0] - 2025-06-07
+
 ### Added
 
 - Added the ability to add timestamps to the start of each line of received data. Timestamps can be enabled from the Settings->RX Settings menu.
@@ -792,7 +794,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added auto-scroll to TX pane, closes #89.
 - Added special delete behaviour for backspace button when in "send on enter" mode, closes #90.
 
-[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.19.1...HEAD
+[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.20.0...HEAD
+[4.20.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.19.1...v4.20.0
 [4.19.1]: https://github.com/gbmhunter/NinjaTerm/compare/v4.19.0...v4.19.1
 [4.19.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.18.0...v4.19.0
 [4.18.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.17.1...v4.18.0